feat: Make A2AClient agnostic to the transport protocol
