+++ client/pom.xml
@@ -29,7 +29,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>a2a-java-sdk-spec-grpc</artifactId>
+            <artifactId>a2a-java-sdk-transport-jsonrpc</artifactId>
             <version>${project.version}</version>
         </dependency>
