+++ reference/grpc/src/main/java/io/a2a/server/grpc/quarkus/QuarkusGrpcHandler.java
@@ -3,7 +3,7 @@ package io.a2a.server.grpc.quarkus;
 import jakarta.enterprise.inject.Instance;
 import jakarta.inject.Inject;
 
-import io.a2a.grpc.handler.GrpcHandler;
+import io.a2a.transport.grpc.server.handler.GrpcHandler;
 import io.a2a.server.PublicAgentCard;
 import io.a2a.server.requesthandlers.CallContextFactory;
 import io.a2a.server.requesthandlers.RequestHandler;
