+++ reference/grpc/src/test/java/io/a2a/server/grpc/quarkus/A2ATestResource.java
@@ -19,7 +19,7 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 
 import io.a2a.server.apps.common.TestUtilsBean;
-import io.a2a.grpc.handler.GrpcHandler;
+import io.a2a.transport.grpc.server.handler.GrpcHandler;
 import io.a2a.spec.PushNotificationConfig;
 import io.a2a.spec.Task;
 import io.a2a.spec.TaskArtifactUpdateEvent;
