+++ transport/jsonrpc/src/main/java/io/a2a/transport/jsonrpc/client/A2AHttpResponse.java
@@ -1,4 +1,4 @@
-package io.a2a.http;
+package io.a2a.transport.jsonrpc.client;
 
 public interface A2AHttpResponse {
     int status();
