+++ transport/jsonrpc/src/main/java/io/a2a/transport/jsonrpc/client/JdkA2AHttpClient.java
@@ -1,4 +1,4 @@
-package io.a2a.http;
+package io.a2a.transport.jsonrpc.client;
 
 import java.io.IOException;
 import java.net.URI;
